homework.10.2-homework.8/homework.10.2-homework.8/project-main/requirements.txt
mongoengine
pika
redis
faker
django-environ